pyproject.toml
[project]
name = "scripts"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "docx>=0.2.4",
    "dotenv>=0.9.9",
    "numpy>=2.3.5",
    "pillow>=12.0.0",
    "python-docx>=1.2.0",
    "vt-py>=0.22.0",
]
